backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
websockets==11.0.3
openai==1.3.5
pydantic==2.4.2
